requirements.txt
coloredlogs
setuptools
numpy
Pillow
